requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database & Auth
supabase==2.3.0
psycopg2-binary==2.9.9

# AI/ML Libraries
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.36.2
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Video Processing
moviepy>=1.0.3
imageio>=2.33.0
imageio-ffmpeg>=0.4.9

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# RAG & Vector Search
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
langchain>=0.1.0
chromadb>=0.4.0

# Web & Streaming
websockets>=12.0
aiofiles>=23.0.0
httpx>=0.24.0,<0.25.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
tqdm>=4.66.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0